Refactor autocomplete
NOTE: This change updates our Node.JS version to 16.4 (current as of this commit) - this
was needed to patch a security vulnerability in one of our dependencies and many other
dependencies are updated as a result, causing a huge diff in package-lock.json - Sorry!

Makes autocomplete work for any TrackedModel without per-model code by standardising the
autocomplete search API endpoints.

All autocompletable models now have a <ModelName>AutoCompleteSerializer, which returns
only the primary-key value and the label to be displayed to the user in search results.
These are the default API endpoints for listing models, since I don't think we use our
API endpoints anywhere currently? We can easily hook these serializers to alternative
endpoints. So a GET query to /api/footnotes?search=foo&format=json will return a JSON
object like this:
```json
[
  {
    "value": "644",
    "label": "TM145 - Foo bar quux",
  },
  {
    "value": "628",
    "label": "TM178 - Bar food flibble",
  },
  ...
]
```

This enables the autocomplete JS to be agnostic of which TrackedModel it is searching
and formatting in results.

The AutoCompleteField custom form field can generate the endpoint URL from the model
class taken from the queryset passed as a constructor argument.