Method that can copy a model
Producing a copy of a model is a common operation
in data work. E.g. it's common to create a new
quota definition for a year by copying the
definition of the previous year and adding a new
validity range. It's also common to copy a
measure and change e.g. the commodity code, to
replicate one set of preferences to another
place.

When copying a measure, it's not enough to just
produce a single record – the components and
conditions have to be replicated as well. This
`copy` method makes doing that easy - it will
find any models within the same TARIC record that
would need to duplicated and creates all of the
new data automatically, applying any specified
overrides.

This is useful enough to warrant sharing between
people doing data migration and may also prove
useful in the user interface at some point.