Duplicate many-to-many fields for `new_versions`
Previously any many-to-many fields that have auto-
generated through models would not be retained by
`new_version`. This commit allows `new_version`
to keep the models if they haven't been overriden
by the user, or set them correctly to a new
iterable.

This won't work for `save=False`, so in the future
we will need to find some way to handle this. It
won't be a problem until we start editing quotas,
which is currently the only place we use
many-to-many fields with an auto-generated through
model.