Auto-select the next SID if one is not supplied
This sets the default value of a SID field to be
computed on demand by the database using a
`SELECT MAX` query.

This means that anyone doing writing of SIDs in an
environment where there could be concurrent writes
should be using table locks to ensure duplicate
SIDs cannot be created.