Ensure locks cannot accidentally apply to global transcations
Locks are *global* to the transaction and once
acquired cannot be released manually. They are
automatically released when the transaction ends.

If `with lock()` were to be called inside another
atomic() block, the lock would not actually be
released at the end of the `with lock()`
statement, which would be a possible source of
hard to spot bugs.

For that reason, this modifies the locking
function to raise `RuntimeError` if the caller is
already in a transaction. This ensures that
either the caller can be sure that the lock is
released when the `with` statement or decorated
function ends, or an error is thrown.