Enable python-xdist with[1] --dist=loadfile[2], this should be cautious enough not to cause issues As the unit per-process is a single file.
Coverage was moved to pytest-cov: pytest-xdist changes the terminal output, which caused
coverage to reset to 0% otherwise.

[1] pytest-xdist distributes tests over mutliple machines or processes, in this case we split
it over processes.

Running locally this can speed up tests (on this developers laptop tests complete in 5 minutes
as opposed to just over 20).

[2] Splitting strategies:  pytest-xdist has different granularites of splitting, from the maximum,
where any test can run anywhere up to file or module as the unit of granularity.

To use the maximum level of granularity, there must be no inadvertantly shared state between tests,
while this is desirable, it doesn't have to be enabled straight away.

loadfile was chosen, as it allows splitting, but won't show up any shared state:  we can do this
as a seperate task, since it can be tricky to find and debug.