TP-193: Create measures (#308)
Creating new measures in the tariff is a complex task, and is broken down into multiple
steps. Measures are made up of a number of linked database models, and so creating a new
measure and all its dependant models must be carefully orchestrated.

This change adds a "wizard", a sequence of connected forms that gather the input data to
construct a new measure and its associated models.

The wizard stores the state of the sequence of forms in the session until the final form
is submitted, and then the models are created, written to the database and linked
together. The user is able to backtrack or even jump around the steps of the sequence,
and is shown a final review page before the measure creation is confirmed.

The sequence of forms is visualized on the right hand column of each step page, using a
step-by-step navigation component.

Some of the steps are forms which allow the user to build up a list of entries, with an
"Add another" button and "Delete" buttons to add and remove entries from the list before
submitting the list and proceeding to the next step.

The geographical area form input is a composite input, where the user can select one of
three options (All countries, a specific country group, or a specific country).
Selecting the group or country option allows the user to make a further selection. There
is currently no way to specify country exclusions from a country group (or all
countries).