Support filtering data to a specific approved transaction
At the moment our implementation of
`approved_up_to_transaction` treats all approved
transactions as a special case, and if you pass
an approved transaction as input to it, it will
implicitly ignore it and return the data as of
the last approved transaction instead.

This was useful whilst we had confusion about
transaction order numbers between the seed file
and subsequent revisions, because it meant that
order didn't matter so much. But it does mean
that we cannot e.g. run a business rule as of
some past approved transaction and have that rule
applied correctly.

This commit modifies the behaviour to now
correctly scope data down to any specific
transaction, even it is from the seed file or has
been approved. This has exposed a number of bugs
in our tests and rules which also have been
fixed.