Fix for QuerySet.select_related error seen in Create Measures
This is a fix for bug TP-947 which was caused by an attempt to
construct a GeographicalAreaQuerySet to include the reverse
related descriptions objects. The underlying mechanism within
TrackedModelQuerySet.with_latest_links() had recently changed
to include reverse relations that could not by applied to
QuerySet.select_related(). This fix prevents reverse relations
being applied to QuerySet.select_related().