Infer authorised use instead of being explicit
Previously uses of the `MeasureCreationPattern`
had to work out whether the measure required
authorised use conditions or not. Invariably this
just ended up as a check on whether the measure
type is one of the authorised use measure types.

Rather than spread this logic around the edges of
the app in multiple places, actually just infer
this in the same way by detecting the presence of
"authorised use" in the measure type's
description. This avoids hard-coding measure
SIDs.