TP-859: Switch to using an in-memory SQLite export
We've found it too difficult to provision enough
temporary disk space on PaaS to create ~3Gb
database. So this commit changes the export to
use an in-memory database instad.

We've changed wrappers to use apsw instead of the
native library so that we can subsequently
deserialize the database into bytes abd strean it
directly to S3.