TP-907 Optional footnotes and conditions (#383)
TP-907 Optional footnotes and conditions
Why
It should be possible to create a measure without any conditions or footnotes. The UI does not permit that at the moment and so this PR enables that capability.

What
Add support in forms to allow empty conditions and footnotes to validate correctly.
Add support in Jinja2 templates to correctly display initial, single form in formsets of conditions and footnotes.
Correctly display empty conditions and footnotes on create measures summary page.
As per new design instructions, add space between the last form in the formset and the “Add new” button to convey to the user that it is not directly associated with the final form.
Add support for removing UI state information from form data when saving to the wizard's session-based backing store.