Add pattern that helps manage suspensions implemented with additional codes
This commit adds a new pattern that can `suspend`
and `unsuspend` rates on a code using the common
pattern of two additional codes. This is used for
policies such as INN, TAPI and UK suspensions for
which there is no longer an appropriate EU code.

Modifying these regimes is difficult to get right
because there are lots of moving parts – one
either needs to delete or restore the original
MFN or not depending on the presence of other
suspensions. There is also a lot of boilerplate
associated with correctly setting up the measure
data.

This code is useful immediately for notebook users
making data changes and may also be useful later
in the UI.