Refactor and fix `in_use` methods to be transaction-aware
Many of our business rules refer to a model being
"in use" by another model, and prevent deleting
the model if this is true. We have a number of
specific `in_use` methods that implement this
check.

Most of these methods are not transaction-aware –
i.e. they operate only on the transaction that
the model was created in. This is a bug, because
it is possible to run the business rule as of a
later transaction – in this case, the check can
return a false negative as it is consdiering an
earlier point in time than was asked for.

This commit removes all of the specific `in_use`
methods and replaces them with a generic version
which can auto-detect the correct relations, and
can also be scoped down to specific relations for
use in business rules.