Fix update validity business rule to consider transactions in order
We have a business rule that is applied to all
models that ensures, per version group, we only
have: 1) the first version being a CREATE, 2)
subsequent versions being an UPDATE or a DELETE,
and 3) no further versions after a DELETE.

At the moment this business rule only expects to
be run against the very latest version, and does
not work when we run it against earlier versions.
This happens routinely when we CREATE and then
UPDATE a model in the same workbasket, because
when we run the rule against the CREATE it picks
up the UPDATE too.

Instead, the business rule needs to be
transaction-aware – i.e. only consider data that
exists as of a certain transaction. This commit
creates a `versions_up_to` method and uses the
updated version ordering from the previous
commit.