Remove assertion on presence of MFN rate
Some codes may not have MFN rates on them. This is
an error in the data but is perfectly permissible
by the system and it meets the business rules.

So it's not right that this is an assertion –
really it is just a warning that the pattern
cannot create an appropriate MFN measure.