TP-1059 Word change on conditions and footnotes
# Why
A recent change allowing users to enter no Condition Codes or Footnotes when creating a Measure meant that the wording on their respective steps did not describe intent accurately to the user. This change updates the wording on those Create Measure steps for Condition Codes and Footnotes and clearly describes optionality.

# What
This PR adds an info element to conditions and footnotes elements of the MeasureCreateWizard.STEPS list. The approach maintains consistency with the intent of that list to provide step-specific information that is accessible within templates. The info element is rendered as informational text on a step only if it is defined (currently only for Condition Codes and Footnotes).

# Checklist
Requires migrations? No
Requires dependency updates? No