Fix effective validity calculation for measures
A measure can have either an explicit or implicit
end date, where the implicit version is specified
by the regulation. Our previous algorithm for
finding the implicit end date was not correct,
because it reported lots of EU measures as still
being in force when clearly (as enforced by ME32)
they could not be.

This commit fixes the algorithm, and also ensures
that when the value is requested from a measure
for which it was not pre-loaded, it is calculated
in the same way.

The new algorithm has been found through
examination of the existing data and comparison
with the Online Tariff service as a known good
source. E.g. I found EU measures that the system
considers still active using:

    Measure.objects
    .with_effective_valid_between()
    .filter(
      sid__lt=20000000, # exclude UK measures
      db_effective_valid_between__contains=today,
    )

With this new algorithm, the above query returns
a small number of measures which it would make
sense if still active e.g. measures on Meursing
codes. There are still some anomalies, but the
situation is much improved.