Return ME67 to active service
At some point in the past, ME67 has been disabled,
probably due to a faulty implementation that was
firing on seed data erroneously.

However, our data does need to abide by ME67 and
we have had a number of recent cases where we've
failed to do so because this rule is disabled.

This commit restores ME67 and adds to its tests to
ensure that it behaves correctly.