Basic tests for views (that they return a 200 and are on the expected url.
While there is a bit of duplication here, it's probably better than having one big test in common.
It's also useful to seperate these from the detail view tests, even if the code is pretty similar.

In future these tests, and the view tests should probably be refactored and expanded to verify
the expected data is included in the output.