TP-758: ME32 without the materialized tree
Our previous implementation of ME32 used the
Nodes contained in the Goods Nomenclature tree to
check for overlaps between commodities. We have
found these Nodes difficult to maintain (it's a
3D-cube of commodity trees!) and subsequently the
data structure is wrong in many places.

This leads to false positives and (presumably) false
negatives from ME32.

This commit introduces a new implementation
of ME32 that does not use the tree and instead
makes use of the in-memory tree used for the
commodity code importer.