TP-935 Transaction-time context for queries
We have a lot of database queries that specify the transaction-time (one of the time
axes of our bitemporal data) to filter results. The most common case is to filter by the
latest approved transaction including draft changes in the current workbasket. This
often results in having to pass the transaction down long function call chains.

This commit adds thread-local storage using Django-CRUM to keep a reference to the
current transaction, so that database querying code can access it directly without
having to pass it around.

We also need to perform queries against different transactions in the same code, and so
we need a way to temporarily override the current transaction.

This is accomplished with a ``contextmanager``, which sets the thread-local stored
transaction to a new value with the scope of a ``with`` block, then resets it
afterwards. Combined with a lazily evaluated query value, this allows defining a
query once and then reusing it with different transactions without having to build a
separate query for each transaction.