TP-1125: Make ME58 consider volumes and units as well
In TP-1020 we made it so that users could not add
more than one condition with the same
certificate. This works for all certificate-based
conditions, but it is not correct for conditions
with volumes (under e.g. condition code E).

ME58 will now stop users from adding conditions
that use volume amounts, because these conditions
don’t have certificates and the rule thinks this
is a duplicate of the base case where nothing is
supplied.

So this commit fixes the rule to only enforce that
the combination of certificate + volume + unit is
unique.