Ensure quotas have a unit or currency
In the TARIC XSD it's possible for neither the
monetary unit or measurement unit to be set. But
this doesn't represent a real situation – the
volume on a quota definition must refer to one of
the two, so one (and only one) of them should
always be present. This is confirmed by the fact
that no previous data has ever had neither or
both present.

This introduces a check constraint that ensures
that only one of the two columns is always
non-null.