Fix importer test that exposes integrity problem in NIG2
For some reason, the importer test for goods
nomenclatures has started failing. It's not clear
to me why that is but as it is on a reverse
relation usage it makes sense that it's due to
the changes made on this branch.

The problem is that the business rule is being run
indirectly against an indent as a result of the
goods nomenclature deletion. The rule reasonably
assumes that the goods nomenclature still exists
– it should be an error to delete any record
without also deleting all of its subrecords in
the same transaction.

However, in all other places the system is
tolerant of this not happening. Really, the
importer test should generate XML that deletes
everything in the same transaction, but it does
not and that's a big change. It's also possible
that EU data will contain this sort of thing, and
it would be better for us to be tolerant of it
rather than failing.

So this change modifies the rule to emit a warning
if the expected goods nomenclature is not present
and quit early instead.