Make sure geo exclusion business rules don't consider deleted memberships
In the previous implementations of ME66 and ON14,
even memberships that had been deleted would be
considered as valid memberships to prevent the
business rule from triggering.

This implementation abstracts out the common logic
in these rules and adds tests that neither will
consider a membership that has been deleted.