Use naming conventions to auto-generate routes
We already have a set of naming conventions that
we are using the UI for views, e.g. "list" is the
page that shows all the things, "edit" is the page
that updates a thing, etc. We use these conventions
already to name our routes and views.

This commit removes the need to endlessly hardcode
our routes by auto-generating the routes based on
what views are present. It will silently ignore
views we don't have yet, so when we add a new view
to do a certain job (e.g. create a certificate) no
more configuration is required beyond just naming
the view class correctly.