TP2000-55: Deletion of all currently viewable models
This commit introduces generic deletion
functionality that allows deletion of objects
from the UI. (Here we mean deletion in the TARIC
sense rather than removing rows from the
database.)

Deletion has been added for all models that
currently have view screens, so for root records
and descriptions, but not commodities.

The actions list has been updated to only show
actions when a URL is actually available. So as we
add edit and delete to more objects, actions for
them will magically appear.

The tariff changes list now also shows what kind
of update is being done to each object – so that
deletes are visibly deleting.

This also shakes up the class hierarchy a bit –
the BusinessRulesMixin is now defined on every
edit view.