TP2000-105: Only persist changes if business rules are passed
This commit ensures that no database changes are
persisted until any business rules are passed as
specified by `validate_business_rules`.

This change is necessary because calling
`new_version` with `save=False` still creates a
transaction, which results in lots of empty
transactions being created. The new behaviour
only calls `new_version` when we are sure that
the model will be saved.