TP2000-77: Run workbasket transitions in background job
We have previously had issues transitioning
workbaskets through states because the transition
would run in a web thread and would time out
after 30ish seconds.

This commit moves those transitions to a
background task so that they can take as long as
necessary and will continue after the lifetime of
the request.