Only EDITING workbaskets should be used for editing
At the moment, any workbasket in a non-approved
state can be selected as the currently active
workbasket. This isn't right because users
shouldn't be editing baskets once they are
PROPOSED and shouldn't be stuck editing a basket
that is ARCHIVED. Essentially, they should only
be editing baskets in the EDITING state.

This commit makes sure that the current workbasket
is in EDITING and if not, picks the next
available EDITING workbasket. If there is none,
it creates one as before.