Ensure that dependent objects are created in earlier transactions
Our testing has been pretty lax on enforcing
transaction order. By default, using SubFactory
will create a new object in a subsequent
transaction, which is often the opposite of what
we want because data referenced by a model needs
to be created before the model.

Then we started creating SubFactories in the same
transactions as the parent model, which mostly
worked except it violates a few assumptions (i.e.
that one transaction only contains one type) and
also leads to problematic XML output.

So this commit changes the scheme so that
dependent subfactories are always created in a
transaction ordered *before* the current one, in
the same partition and workbasket. It doesn't
really matter that lots of transactions end up
with the same order, as long as they order
correctly.