Make populating the validity field a queryset responsibility
Previously there was `objects_with_validity_field`
which would ensure objects were loaded with their
validity field present. This is necessary because
for some models (Measures and Regulations) their
validity is implicit and so has to be computed by
a specific algorithm.

It was done this way because at the time there was
no `ValidityQuerySet` and we didn't want to
introduce one. Since then though one has
sprouted, and it is clearly a more natural place
for this interface to sit.

So this commit refactors that interface to turn it
into `queryset.with_validity_field()` which seems
to fit the usage much better anyway.