Utilities for finding a set of current related models
This commit adds `follow_path` which when given a
model and a double-underscore-separated path of
relations to follow, will return a queryset
containing the related objects at the end of the
path, ensuring that any objects "passed through"
are also current and not deleted.

This is useful for concisely describing relations
and resolving them, which will be used to reduce
a lot of complexity in business rules that
examine related objects. It will also be useful
when we implement "searching through" as it will
allow the searches to be defined as simple path
strings that the system can then resolve into a
filterable queryset.