TP2000-166: Improve performance of validity rules
It turns out the recent changes to validity rules
are not very performant when there are lots of
joins involved, because the database has to
compute validity dates for nearly all measures
and then filter them to just the one or two are
interested in.

Because the number of objects we expect to return
from business rules is quite small, instead we
materialize the result set for each join and just
dispatch them as separate queries. This seems to
be faster that allowing the database to do that.