Document training scenarios
Our users need practice on how to use our tool so
that they can operate independently in the
future.

This commit adds a number of training scenarios in
BDD format (to allow future automation). When our
documentation site is built, these are converted
into HTML and included as part of our normal
documentation.