Remove test of transaction sort order
This tests that the default ordering that Django
applies to transactions is not the same as the
default ordering that Psotgres will return
transactions in if no ordering is specified.

This test doesn't make any sense because if no
ordering is specified to Postgres it is free to
return objects in _any_ order, which includes the
one being tested for. Hence at best, this test
will be flaky and worst it will just always be
wrong. This is not a useful test so I'm removing
it.