Run imports inside a single database transaction
We have recently had issues that have prevented
our imports from completing successfully. In
these cases we've been left with the database in
a inconsistent state and in order to reattempt
the import we have to restore the database from a
backup.

As discussed, this commit changes the behaviour so
that now an entire import is run inside a single
database transaction. If there is an error that
causes the import to bomb out, the changes are
removed, and the issue can be fixed and the
import reattempted more easily.

This applies to imports at the "chunk" level. Most
import jobs will only generate one chunk, so that
is equivalent to a single file. For imports that
span across multiple independent chunks, each
chunk will have to be manually restarted.