Implement ME27 for regulation replacements
We never implemented this rule because we thought
we weren't using replacements in the UK. We're
still not, but if a user (accidentally?) selects
an EU regulation it's possible for the rule to be
broken. Selecting EU regulations is still
necessary because the UK has "retained" some EU
legislation and it is still valid to choose it.