TP2000-206: Automatically select terminating regulations
At the moment it's not possible for users to add
or remove end dates to measures because they
aren't able to make a corresponding edit to the
terminating regulation. This behaviour is also
handled in a number of bespoke places in the code
base.

This commit unifies that logic into a signal that
is fired pre-save, so no matter where the edit to
the measure is happening the logic will be
applied. This also lets us remove some redundant
code.

The ``mute_signals`` decorator is necessary for
ME33 and ME34 so that it _is_ possible to create
test data that ignore this logic.