Tp2000 278  WorkBasket list and detail views (#549)
TP2000-278 WorkBasket list and detail views
Why

There's no provision to allow WorkBaskets, other than the one that is currently in editing state, to be viewed and reviewed. This is a problem both for:

    Managing the operations process, which requires tariff manages and engineers accurately communicating WorkBasket status.
    QA and peer review of changes in non-editing workbaskets

What

Added workbasket list view with specific workbasket selection to view details. The details view provides top-level info to assist with operations tasks (workbasket state changes, exports, etc) and the changes assists QA and peer reviewing.

    Requires migrations? No
    Requires dependency updates? No