Models for open data simplified database (#1423)
There is a desire to use TAP data in open data analysis. The data currently published contain all the historical information, and analysis are complex.

This PR create a set of 'current' data that can be exported every day and used to create reports within TAP.

It creates a shadow model of each tracked model in TAP. The shadow model contains only the approved and current version.
Where feasible, models have been merged: for instance, descriptions are in the described model, not in a different one.

The tables supporting the models are created in a different partition of the database, so they can easily be excluded from the daily back up of the database.

There is a mechanism to populate the data : this will eventually happen daily, but it is not part of the current PR.
Some raw sql is used to copy the data (to improve performance). Even so, the copy process on localhost takes a couple of hours. This may or may not be a problem, it is too early to know.

Requires migrations: Yes
Requires dependency updates - No